Refactor achievement organization to by-mode structure
Reorganized achievement files to follow a consistent by-mode pattern
where all achievements for each game mode are in a single file.

Changes:
- Created osu.py consolidating all osu!std achievements (32 total)
  - Skill pass/FC (20)
  - Combo milestones (4)
  - Playcount milestones (4)
  - Rank milestones (4)

- Added rank achievements to mode files:
  - taiko.py: 20 → 24 achievements
  - catch.py: 20 → 24 achievements
  - mania.py: 20 → 24 achievements

- Added section comments to all mode files for clarity:
  - Skill Pass Achievements
  - Skill FC Achievements
  - Hit Count Achievements (taiko/catch/mania)
  - Combo Achievements (osu)
  - Playcount Achievements (osu)
  - Rank Achievements

- Removed old split files:
  - osu_skill.py
  - osu_combo.py
  - osu_plays.py
  - rank_milestones.py

- Updated app/achievements/__init__.py registry imports

Total achievements remain 128 (verified):
- osu.py: 32
- taiko.py: 24
- catch.py: 24
- mania.py: 24
- mod_introductions.py: 11
- hush_hush.py: 13
